@@ -42,7 +42,7 @@
     "sinon": "^7.3.1"
   },
   "engines": {
-    "node": ">= 8.0.0"
+    "node": ">= 10.0.0"
   },
   "lint-staged": {
     "*.{js,json,md}": [
