@@ -1,6 +1,6 @@
 {
   "name": "@parcel/cache",
-  "version": "2.0.0-alpha.1.1",
+  "version": "2.0.0-alpha.2.1",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -8,10 +8,10 @@
   },
   "main": "src/Cache",
   "engines": {
-    "node": ">= 8.0.0"
+    "node": ">= 10.0.0"
   },
   "dependencies": {
-    "@parcel/logger": "^2.0.0-alpha.1.1",
-    "@parcel/utils": "^2.0.0-alpha.1.1"
+    "@parcel/logger": "^2.0.0-alpha.2.1",
+    "@parcel/utils": "^2.0.0-alpha.2.1"
   }
 }
