@@ -1,6 +1,6 @@
 {
   "name": "@parcel/logger",
-  "version": "2.0.0-alpha.1.1",
+  "version": "2.0.0-alpha.2.1",
   "description": "Blazing fast, zero configuration web application bundler",
   "main": "src/Logger.js",
   "license": "MIT",
@@ -9,12 +9,12 @@
     "url": "https://github.com/parcel-bundler/parcel.git"
   },
   "engines": {
-    "node": ">= 8.0.0"
+    "node": ">= 10.0.0"
   },
   "publishConfig": {
     "access": "public"
   },
   "dependencies": {
-    "@parcel/events": "^2.0.0-alpha.1.1"
+    "@parcel/events": "^2.0.0-alpha.2.1"
   }
 }
