@@ -1,2 +1,2 @@
 import fs from 'fs';
-module.exports = fs.readFileSync(__dirname + '/test.txt', 'utf8');
+export default fs.readFileSync(__dirname + '/test.txt', 'utf8');
