@@ -5,7 +5,7 @@ import type {
   ModuleSpecifier,
   Symbol
 } from '@parcel/types';
-import {md5FromString} from '@parcel/utils';
+import {md5FromObject} from '@parcel/utils';
 import type {Dependency, Environment, Target} from './types';
 
 type DependencyOpts = {|
@@ -22,17 +22,20 @@ type DependencyOpts = {|
   env: Environment,
   meta?: Meta,
   target?: Target,
-  symbols?: Map<Symbol, Symbol>
+  symbols?: Map<Symbol, Symbol>,
+  pipeline?: ?string
 |};
 
 export function createDependency(opts: DependencyOpts): Dependency {
   let id =
     opts.id ||
-    md5FromString(
-      `${opts.sourceAssetId ?? ''}:${opts.moduleSpecifier}:${JSON.stringify(
-        opts.env
-      )}:${opts.target ? JSON.stringify(opts.target) : ''}`
-    );
+    md5FromObject({
+      sourceAssetId: opts.sourceAssetId,
+      moduleSpecifier: opts.moduleSpecifier,
+      env: opts.env,
+      target: opts.target,
+      pipeline: opts.pipeline
+    });
 
   return {
     ...opts,
