@@ -21,13 +21,13 @@ import SourceMap from '@parcel/source-map';
 import {
   bufferStream,
   loadConfig,
-  md5FromFilePath,
   md5FromString,
   blobToStream,
   TapStream
 } from '@parcel/utils';
 import {createDependency, mergeDependencies} from './Dependency';
 import {mergeEnvironments} from './Environment';
+import {PARCEL_VERSION} from './constants';
 
 type AssetOptions = {|
   id?: string,
@@ -139,7 +139,7 @@ export default class InternalAsset {
     // and hash while it's being written to the cache.
     let [contentKey, mapKey] = await Promise.all([
       this.options.cache.setStream(
-        this.generateCacheKey('content' + pipelineKey),
+        this.getCacheKey('content' + pipelineKey),
         contentStream.pipe(
           new TapStream(buf => {
             size += buf.length;
@@ -150,7 +150,7 @@ export default class InternalAsset {
       this.map == null
         ? Promise.resolve()
         : this.options.cache.set(
-            this.generateCacheKey('map' + pipelineKey),
+            this.getCacheKey('map' + pipelineKey),
             this.map
           )
     ]);
@@ -219,8 +219,10 @@ export default class InternalAsset {
     this.map = map;
   }
 
-  generateCacheKey(key: string): string {
-    return md5FromString(key + this.value.id + (this.value.hash || ''));
+  getCacheKey(key: string): string {
+    return md5FromString(
+      PARCEL_VERSION + key + this.value.id + (this.value.hash || '')
+    );
   }
 
   addDependency(opts: DependencyOptions) {
@@ -241,11 +243,7 @@ export default class InternalAsset {
     return dep.id;
   }
 
-  async addIncludedFile(file: File) {
-    if (file.hash == null) {
-      file.hash = await md5FromFilePath(this.options.inputFS, file.filePath);
-    }
-
+  addIncludedFile(file: File) {
     this.value.includedFiles.set(file.filePath, file);
   }
 
