@@ -1,8 +1,10 @@
+import * as other from './other';
+
 type Params = {
-  hello: string;
+  hello: other.Params;
 };
 
-export class Test {}
+// export class Test {}
 
 export default function test(params: Params) {
   return params.hello;
@@ -14,4 +16,8 @@ export function foo() {
 
 var x = 2;
 var p = x + 2, q = 3;
-export {p};
+export {p as hi};
+// export {Test as Hello} from './other';
+// export * from './other';
+
+export {default as a} from './other';
